+++ db/migrations/20240601-create-referral-bonuses.js
@@ -32,7 +32,7 @@ module.exports = {
         allowNull: false,
       },
       amount: {
-        type: Sequelize.DECIMAL(18, 8),
+        type: Sequelize.DOUBLE,
         allowNull: false,
       },
       createdAt: {
